Map Local key to Windows/Command
When using the USB converter with a Mac, the lack of a mapping for the
Command key becomes a real issue.  With this change, the Local key is
mapped as Command (Windows) key.  The version report key (Local-V) is
gone, the boot loader jump function is now mapped to Local+Abort.

The build process now builds both Teensy 1 and Teensy 2 versions of the
firmware.  Both hex files are stored in GitHub.

The timing has been slightly changed again, but there are still some
issues which cause unwanted keys to be pressed.  Any reports regarding
this and possible timing fixes would be welcome.